@HyukjinKwon 2018-08-03
Merged to master.